Fixed recyclerview, added play/pause buttons, imported song picture urls
