+++ app/build.gradle
@@ -33,25 +33,20 @@ dependencies {
     })
 
     // This library handles authentication and authorization
-    
-
     // This library handles music playback
-    
-    
+
     compile 'com.android.support:appcompat-v7:25.3.1'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
     compile 'com.spotify.sdk:spotify-android-auth:1.0.0@aar'
     compile 'com.spotify.sdk:spotify-player:24-noconnect-2.20b@aar'
     compile 'com.google.firebase:firebase-core:11.0.1'
     compile 'com.google.firebase:firebase-auth:10.0.1'
-
     compile 'com.firebaseui:firebase-ui-auth:2.0.1'
-    compile('com.facebook.android:facebook-android-sdk:[4,5)')
-
+    compile 'com.facebook.android:facebook-android-sdk:[4,5)'
     compile 'com.android.support:recyclerview-v7:25.3.1'
-
     testCompile 'junit:junit:4.12'
 }
+
 
 
 apply plugin: 'com.google.gms.google-services'
