+++ app/src/main/java/com/studios/juke/juke/GenerateQrActivity.java
@@ -1,0 +1,19 @@
+package com.studios.juke.juke;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.support.v7.widget.Toolbar;
+
+public class GenerateQrActivity extends MenuBarOptions {
+
+    private Toolbar mToolbar;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_generate_qr);
+
+        mToolbar = (Toolbar) findViewById(R.id.my_toolbar);
+        setSupportActionBar(mToolbar);
+    }
+}
