+++ app/src/main/java/com/studios/juke/juke/MainActivity.java
@@ -1,295 +1,29 @@
 package com.studios.juke.juke;
 
-import android.app.Activity;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 
-import android.widget.EditText;
-import android.widget.Toast;
+/**
+ * Created by bdo04 on 6/25/2017.
+ */
 
-import com.spotify.sdk.android.authentication.AuthenticationClient;
-import com.spotify.sdk.android.authentication.AuthenticationRequest;
-import com.spotify.sdk.android.authentication.AuthenticationResponse;
-import com.spotify.sdk.android.player.Config;
-import com.spotify.sdk.android.player.ConnectionStateCallback;
-import com.spotify.sdk.android.player.Error;
-import com.spotify.sdk.android.player.Player;
-import com.spotify.sdk.android.player.PlayerEvent;
-import com.spotify.sdk.android.player.Spotify;
-import com.spotify.sdk.android.player.SpotifyPlayer;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-
-import java.util.ArrayList;
-
-import static android.content.ContentValues.TAG;
-
-public class MainActivity extends AppCompatActivity implements SpotifyPlayer.NotificationCallback, ConnectionStateCallback {
-
-    private EditText mEditSong;
-    private Button mSearchButton;
-    private Toolbar mToolbar;
-
-    // TODO: Replace with your client ID
-    private static final String CLIENT_ID = "256fa987714a455687888ed1f07c3630";
-    // TODO: Replace with your redirect URI
-    private static final String REDIRECT_URI = "juke-login://callback";
-    // Request code that will be used to verify if the result comes from correct activity
-    // Can be any integer
-    private static final int REQUEST_CODE = 1337;
-
-    //dont leave this static
-    public static Player mPlayer;
-    private String mAccessToken;
-    private ArrayList<Song> mSongs = new ArrayList<>();
+public class MainActivity extends AppCompatActivity {
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState){
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        Button start_button = (Button)findViewById(R.id.start);
 
-        mToolbar = (Toolbar) findViewById(R.id.my_toolbar);
-        setSupportActionBar(mToolbar);
-
-        mEditSong = (EditText) findViewById(R.id.edit_song);
-        mSearchButton = (Button) findViewById(R.id.search_button);
-
-       // pauseButton.setTag(1);
-        mSearchButton.setOnClickListener(new View.OnClickListener() {
+        start_button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                String searched_song = mEditSong.getText().toString();
-                new GetSpotifyTracks(searched_song).execute();
-
-
-                /*final int status = (Integer) v.getTag();
-                if (status == 1) {
-                    mPlayer.pause(null);
-                    pauseButton.setText("Play");
-                    v.setTag(0);
-                } else {
-                    mPlayer.resume(null);
-                    pauseButton.setText("Pause");
-                    v.setTag(1);
-                }*/
+                Intent intent = new Intent(MainActivity.this,SpotifyHomeActivity.class);
+                startActivity(intent);
             }
         });
-
-        AuthenticationRequest.Builder builder = new AuthenticationRequest.Builder(CLIENT_ID,
-                AuthenticationResponse.Type.TOKEN,
-                REDIRECT_URI);
-        builder.setScopes(new String[]{"user-read-private", "streaming"});
-        AuthenticationRequest request = builder.build();
-
-        AuthenticationClient.openLoginActivity(this, REQUEST_CODE, request);
-
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        super.onCreateOptionsMenu(menu);
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.tool_bar, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_settings:
-                return true;
-
-            case R.id.action_home:
-                return true;
-
-            case R.id.action_search:
-                return true;
-
-            default:
-                // If we got here, the user's action was not recognized.
-                // Invoke the superclass to handle it.
-                return super.onOptionsItemSelected(item);
-
-        }
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
-        super.onActivityResult(requestCode, resultCode, intent);
-
-        // Check if result comes from the correct activity
-        if (requestCode == REQUEST_CODE) {
-            AuthenticationResponse response = AuthenticationClient.getResponse(resultCode, intent);
-            if (response.getType() == AuthenticationResponse.Type.TOKEN) {
-                mAccessToken = response.getAccessToken();
-                Log.i("Access Token: ", mAccessToken);
-                Config playerConfig = new Config(this, mAccessToken, CLIENT_ID);
-                Spotify.getPlayer(playerConfig, this, new SpotifyPlayer.InitializationObserver() {
-                    @Override
-                    public void onInitialized(SpotifyPlayer spotifyPlayer) {
-                        mPlayer = spotifyPlayer;
-                        mPlayer.addConnectionStateCallback(MainActivity.this);
-                        mPlayer.addNotificationCallback(MainActivity.this);
-                    }
-
-                    @Override
-                    public void onError(Throwable throwable) {
-                        Log.e("MainActivity", "Could not initialize player: " + throwable.getMessage());
-                    }
-                });
-            }
-        }
-    }
-
-    @Override
-    protected void onDestroy() {
-        //dispose of the Player when we are done with it (NECESSARY!)
-        Spotify.destroyPlayer(this);
-        super.onDestroy();
-    }
-
-    @Override
-    public void onPlaybackEvent(PlayerEvent playerEvent) {
-        Log.d("MainActivity", "Playback event received: " + playerEvent.name());
-        switch (playerEvent) {
-            // Handle event type as necessary
-            default:
-                break;
-        }
-    }
-
-    @Override
-    public void onPlaybackError(Error error) {
-        Log.d("MainActivity", "Playback error received: " + error.name());
-        switch (error) {
-            // Handle error type as necessary
-            default:
-                break;
-        }
-    }
-
-    @Override
-    public void onLoggedIn() {
-        Log.d("MainActivity", "User logged in");
-    }
-
-    @Override
-    public void onLoggedOut() {
-        Log.d("MainActivity", "User logged out");
-    }
-
-    @Override
-    public void onLoginFailed(Error error) {
-        Log.d("MainActivity", "Login failed");
-    }
-
-    @Override
-    public void onTemporaryError() {
-        Log.d("MainActivity", "Temporary error occurred");
-    }
-
-    @Override
-    public void onConnectionMessage(String message) {
-        Log.d("MainActivity", "Received connection message: " + message);
-    }
-
-    private class GetSpotifyTracks extends AsyncTask<Void, Song, Void> {
-
-        String mSearchedSong;
-
-
-        GetSpotifyTracks(String searched_song){
-            mSearchedSong = searched_song;
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            Toast.makeText(MainActivity.this,"Json Data is downloading",Toast.LENGTH_LONG).show();
-        }
-
-        @Override
-        protected Void doInBackground(Void... arg0) {
-            HttpHandler sh = new HttpHandler();
-            // Making a request to url and getting response
-            String url = "https://api.spotify.com/v1/search?q=" + mSearchedSong + "&type=track";
-            String jsonStr = sh.makeServiceCall(url, mAccessToken);
-
-            Log.e(TAG, "Response from url: " + jsonStr);
-            if (jsonStr != null) {
-                try {
-                    JSONObject root = new JSONObject(jsonStr);
-
-                    // Getting JSON Array node
-                    JSONObject tracks = root.getJSONObject("tracks");
-                    JSONArray items = tracks.getJSONArray("items");
-
-                    for(int i = 0; i < items.length(); ++i) {
-                        JSONObject track = items.getJSONObject(i);
-                        String song_name = track.getString("name");
-                        JSONArray artists = track.getJSONArray("artists");
-                        JSONObject main_artist = artists.getJSONObject(0);
-                        String artist = main_artist.getString("name");
-                        String uri = track.getString("uri");
-
-                        Song song = new Song(song_name, artist, uri);
-
-                        publishProgress(song);
-                    }
-
-                } catch (final JSONException e) {
-                    Log.e(TAG, "Json parsing error: " + e.getMessage());
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            Toast.makeText(getApplicationContext(),
-                                    "Json parsing error: " + e.getMessage(),
-                                    Toast.LENGTH_LONG).show();
-                        }
-                    });
-
-                }
-
-            } else {
-                Log.e(TAG, "Couldn't get json from server.");
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        Toast.makeText(getApplicationContext(),
-                                "Couldn't get json from server. Check LogCat for possible errors!",
-                                Toast.LENGTH_LONG).show();
-                    }
-                });
-            }
-
-            return null;
-        }
-
-        @Override
-        protected void onProgressUpdate(Song... songs) {
-            super.onProgressUpdate(songs);
-            mSongs.add(songs[0]);
-        }
-
-        @Override
-        protected void onPostExecute(Void result) {
-            super.onPostExecute(result);
-
-            //After all songs are queried.. THEN start intent!
-            Intent intent = SongListActivity.newIntent(MainActivity.this, mSongs);
-            startActivity(intent);
-        }
     }
 }
