+++ app/src/main/java/com/studios/juke/juke/MenuBarOptions.java
@@ -1,0 +1,52 @@
+package com.studios.juke.juke;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.util.Log;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+
+public class MenuBarOptions extends AppCompatActivity {
+
+    private Toolbar mToolbar;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState){
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        super.onCreateOptionsMenu(menu);
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.menu_bar, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.action_settings:
+                return true;
+
+            case R.id.action_home:
+                Intent intent1 = new Intent(MenuBarOptions.this, SpotifyHomeActivity.class);
+                startActivity(intent1);
+                return true;
+
+            case R.id.action_search:
+                Intent intent2 = new Intent(MenuBarOptions.this, SpotifySearchActivity.class);
+                startActivity(intent2);
+                return true;
+
+            default:
+                // If we got here, the user's action was not recognized.
+                // Invoke the superclass to handle it.
+                return super.onOptionsItemSelected(item);
+
+        }
+    }
+}
