+++ app/src/main/res/layout/activity_fragment.xml
@@ -5,5 +5,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent" >
 
+    <include
+        android:id="@+id/my_toolbar"
+        layout="@layout/common_toolbar"/>
+
+
 </FrameLayout>
