+++ app/src/main/res/layout/activity_generate_qr.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="#ED5D67"
+    android:orientation="horizontal"
+    tools:context="com.studios.juke.juke.GenerateQrActivity">
+
+    <include
+        android:id="@+id/my_toolbar"
+        layout="@layout/common_toolbar"/>
+
+
+</android.support.constraint.ConstraintLayout>
