+++ app/src/main/res/layout/list_item_song.xml
@@ -19,11 +19,11 @@
         android:layout_marginStart="8dp"
         android:layout_marginTop="16dp"
         android:text="Song Title"
-        android:textColor="@android:color/black"
-        android:textSize="18sp"
+        android:textColor="@android:color/background_dark"
+        android:textSize="24sp"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toLeftOf="@+id/crime_solved"
+        app:layout_constraintRight_toLeftOf="@+id/song_picture"
         app:layout_constraintTop_toTopOf="parent"/>
 
     <TextView
@@ -36,20 +36,36 @@
         android:layout_marginStart="8dp"
         android:layout_marginTop="8dp"
         android:text="Artist"
+        android:textSize="18sp"
         app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toLeftOf="@+id/crime_solved"
-        app:layout_constraintTop_toBottomOf="@+id/song_title"/>
+        app:layout_constraintRight_toLeftOf="@+id/song_picture"
+        app:layout_constraintTop_toBottomOf="@+id/song_title"
+        app:layout_constraintHorizontal_bias="0.0"/>
 
     <ImageView
-        android:id="@+id/crime_solved"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="16dp"
+        android:id="@+id/song_picture"
+        android:layout_width="179dp"
+        android:layout_height="183dp"
         android:layout_marginRight="16dp"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
         android:layout_marginTop="16dp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"/>
-        <!--app:srcCompat="@drawable/ic_solved"-->
+        android:layout_marginBottom="8dp"
+        app:layout_constraintVertical_bias="0.0"
+        />
+
+    <ImageView
+        android:id="@+id/play_button"
+        android:layout_width="60dp"
+        android:layout_height="60dp"
+        app:srcCompat="@drawable/ic_play_arrow_black_48dp"
+        android:layout_marginLeft="8dp"
+        app:layout_constraintLeft_toLeftOf="parent"
+        android:layout_marginTop="47dp"
+        app:layout_constraintTop_toBottomOf="@+id/song_artist"
+        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginBottom="8dp"
+        app:layout_constraintVertical_bias="1.0"/>
 
 </android.support.constraint.ConstraintLayout>
